auction-house/requirements.txt
google-cloud-bigquery==1.27.2